+++ src/constants/messages.ts
@@ -1,0 +1,7 @@
+export enum messages {
+  EmailRequired = 'Email is required',
+  EmailInvalid = 'Invalid email address',
+  PasswordRequired = 'Password is required',
+  PasswordMinLength = 'Password must be at least 8 characters long',
+  LoginSucces = 'Your email address has been successfully saved in the local storage!',
+}
